Move lale dataframe continue
